requirements/requirements.txt
pydantic==1.10.2
datetime==4.7
PyYAML==6.0.0
pandas==1.5.2
